requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
plotly==5.24.0
psutil==6.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
retrying==1.3.4
setuptools==72.1.0
six==1.16.0
speedtest-cli==2.1.3
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.43.0
zipp==3.20.1
